+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* `Vector2<T>` types with `Vec2` and `IVec2` aliases
 * Support `no_std` (by disabling the `std` feature flag)